Add files via upload
Mean.py, a common statistical method, or metric.  Using the metric or method, applied as a function within the PeriodMetrics (syntax:  Name of Class proposed) UDO, there is some value I generating the result of an average, moving average, simple or otherwise.  We look at the price in relation to a moving average for signals of a stock's health. if above the average or average over time (moving), then we are able to suggest there are opportunities for investors to continue looking at this one closely.  Inversely, not the best of looks to consider.  Start point on this function is define a list with the ETF/Stock's high over the last 52 weeks, and its present price.  When adding a layer of complexity to the analysis, a list becomes unwieldy, so in effect the list can be transformed into the dictionary data type, simply because, I can look at the numbers of that list or values with specific reference points; if I know the price I am looking at is not special, eg., it is just a price on a given day at a given time, then a list is ok to use, but when looking at a value of considerable known import, eg., 52 week, high, a dictionary will allow one to associate the price on that given day and time with an index value or a flag, namely the key value (eg., hi: 234).  If I take the average of the high over last year and present price I can see the general points of interest or begin to divulge.  One may look at the straight line average from the day and time of the high and the price and compare how the values since plotted against that straight line.  Used in conjunction with a one year moving average, SMA, then we can see the distinction, sharply falling or falling, sharply rising or rising.  In the case of last year or two we know this was likely a bearish place to examine.  If we can see flattening, we can see some directional movement or momentum junctions.  In python we can gather data files from a remote location for analysis and pick and pull the items such as those key points mentioned above, high, current positioning them in a data type apart from that main extraction (a list of daily values over last x periods -- eg., one year), so the dictionary then contains the info we are interested in.  Pulling that info allows further extraction for analysis.  So pulling the values from key points, eg., high and current, we can take an average from that pulled/extracted values list.  These multip step or multi stage functions lend themselves to Ci/Cd and ETl pipelines which orchestration is worth noting.  DevOps engineering applied(s).  .